fix "About" URL in CustomHeader.vue
NOTE: I have not tested this change!

When I am on https://mappingnorthkorea.com/map and click the question mark in the header row a modal opens. I believe the content of that modal is defined here.

When I click "Read more on the about page here." I am redirected to the URL https://mappingnorthkorea.com/map#/about and see only the map, but not the about page.

When I click in the navigation on About I am redirected to the URL https://mappingnorthkorea.com/about and can read the content I'm interested in. Therefore I assume that the link in this file needs to be changed.